removed warnings and segfaults

+++ src/BaseInfo.cpp
@@ -100,12 +100,12 @@ int BaseInfo::CountProductionBuildings()
 		ai->utility->RemoveGraphics(basePerimiterDrawID);
 	}
 
-	for (int c = 0; c < clusters.size(); c++ )
+	for (int c = 0; c < (int)clusters.size(); c++ )
 	{	
 		if (clusters[c].size()==2) {
 			basePerimiterDrawID = ai->utility->DrawCircle(clusters[c][0], 70, basePerimiterDrawID);
 		} else {
-			for (int i = 0; i < clusters[c].size() -1 ; i++ )
+			for (int i = 0; i < (int)clusters[c].size() -1 ; i++ )
 			{
 				basePerimiterDrawID = ai->utility->DrawLine(  clusters[c][i], clusters[c][i+1], true, 20, basePerimiterDrawID );
 			}
